requirements.txt
fastapi
uvicorn[standard]
stripe
requests
python-dotenv